requirements.txt
requests~=2.32.3
darkdetect~=0.8.0
pywinstyles~=1.8
python-decouple~=3.8